fix final term grade

+++ student_records/mysql_db/init_db.sql
@@ -185,6 +185,6 @@ CREATE TABLE student_final_term_grade
     id_course INT NOT NULL,
     period_term INT NOT NULL,
     period_year INT NOT NULL,
-    period_grade INT NOT NULL,
+    period_grade FLOAT NOT NULL,
     PRIMARY KEY(id_student,id_class,id_course,period_term,period_year)
 );
